Build on msvc x86-64.

+++ components/render/manager/include/render/manager/manager.h
@@ -105,8 +105,8 @@ class RenderManager
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 ///Создание целей рендеринга
 ///////////////////////////////////////////////////////////////////////////////////////////////////
-    RenderTarget CreateRenderBuffer       (size_t width, size_t height, PixelFormat format);
-    RenderTarget CreateDepthStencilBuffer (size_t width, size_t height);
+    RenderTarget CreateRenderBuffer       (unsigned int width, unsigned int height, PixelFormat format);
+    RenderTarget CreateDepthStencilBuffer (unsigned int width, unsigned int height);
                                                          
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 ///Создание примитивов
@@ -121,9 +121,9 @@ class RenderManager
     Texture          CreateTexture          (const media::Image& image, TextureDimension dimension, bool generate_mipmaps = true);
     Texture          CreateTexture          (const media::CompressedImage& image);
     Texture          CreateTexture          (const media::CompressedImage& image, TextureDimension dimension);
-    Texture          CreateTexture2D        (size_t width, size_t height, PixelFormat format, bool generate_mipmaps = true);
-    Texture          CreateTexture3D        (size_t width, size_t height, size_t depth, PixelFormat format, bool generate_mipmaps = true);
-    Texture          CreateTextureCubemap   (size_t size, PixelFormat format, bool generate_mipmaps = true);
+    Texture          CreateTexture2D        (unsigned int width, unsigned int height, PixelFormat format, bool generate_mipmaps = true);
+    Texture          CreateTexture3D        (unsigned int width, unsigned int height, unsigned int depth, PixelFormat format, bool generate_mipmaps = true);
+    Texture          CreateTextureCubemap   (unsigned int size, PixelFormat format, bool generate_mipmaps = true);
     Material         CreateMaterial         ();
     
 ///////////////////////////////////////////////////////////////////////////////////////////////////
