+++ components/render/manager/sources/texture_manager.h
@@ -18,7 +18,7 @@ class TextureManager: public Object
     TexturePtr CreateTexture (const media::Image& image, TextureDimension dimension, bool generate_mipmaps, const char* name = "");
     TexturePtr CreateTexture (const media::CompressedImage& image, const char* name = "");
     TexturePtr CreateTexture (const media::CompressedImage& image, TextureDimension dimension, const char* name = "");
-    TexturePtr CreateTexture (TextureDimension dimension, size_t width, size_t height, size_t depth, PixelFormat format, bool generate_mipmaps, const char* name = "");
+    TexturePtr CreateTexture (TextureDimension dimension, unsigned int width, unsigned int height, unsigned int depth, PixelFormat format, bool generate_mipmaps, const char* name = "");
     
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 ///Проверка: является ли ресурс текстурой
